Make PXE actually fully work.
PXE with ProxyDHCP is extremely subtle and quick to anger. The only
reliable way I've found of getting it to boot is to serve a full-blown
boot menu over PXE's terrible protocol, and configure it such that
the ROM autoboots the first entry.

This requires another listening port to (finally!) provide the boot file
location, but works reliably on my Realtek PXE ROMs, as well as ipxe.