Support buggy PXE ROMs that don't send a client GUID.
There is at least one PXE ROM in the world that does this (see #21).
It's a minor deviation from the spec, but it shouldn't result in
any bad behavior.

Also created a new README section to document deliberate deviations
from the spec, because if more of these kinds of things come up,
eventually I'm going to forget why we don't exactly comply.